Extracted side videos into own view component
